% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{.normalize_channels}
\alias{.normalize_channels}
\title{Clean Up channels parameter}
\usage{
.normalize_channels(channels)
}
\arguments{
\item{channels}{An optional character vector of channel names or ids.}
}
\value{
A character vector of channel ids, with the channel names as names of
the vector.
}
\description{
Clean Up channels parameter
}
\keyword{internal}
