% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_all_channel_conversations}
\alias{fetch_all_channel_conversations}
\title{Fetch Channel Conversations}
\usage{
fetch_all_channel_conversations(channels)
}
\arguments{
\item{channels}{An optional character vector of channel names or ids.}
}
\value{
A named list of the \code{conversations.history} objects returned by
\link[slackthreads:conversations]{slackthreads::conversations}, one for each channel. The names of are the
channel names. The list has additional class \code{multichannel.history}.
}
\description{
Fetch all conversations that are available for the given channels. If
channels are not provided, all conversations for all non-archived channels
are fetched.
}
