% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_all_conversation_replies}
\alias{fetch_all_conversation_replies}
\title{Fetch Replies for Everything}
\usage{
fetch_all_conversation_replies(conversations)
}
\arguments{
\item{conversations}{An object returned by
\code{\link[=fetch_all_channel_conversations]{fetch_all_channel_conversations()}}, or (soon) a list with a similar
structure. And I'll probably allow a tidy conversation tibble before I
consider this package done.}
}
\value{
A list of the objects returned by
\code{\link[slackthreads:all_conversation_replies]{slackthreads::all_conversation_replies()}}. The names of the list are the
channel from which the conversations were fetched, and the object has
additional class \code{multichannel.replies}.
}
\description{
Fetch replies for all conversations in an object returned by
\code{\link[=fetch_all_channel_conversations]{fetch_all_channel_conversations()}}.
}
