% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_conversations.R
\name{tidy_conversations}
\alias{tidy_conversations}
\title{Rectangle Conversation Data}
\usage{
tidy_conversations(conversations, threads = NULL)
}
\arguments{
\item{conversations}{A list returned by \code{\link[slackthreads:conversations]{slackthreads::conversations()}}.
Right now we actually are working with a list of such lists (one list per
channel). I'll likely add methods to this to deal with that.}

\item{threads}{An optional list the same shape as conversations, with replies
to those conversations.}
}
\value{
A tibble of question data.
}
\description{
Rectangle Conversation Data
}
