% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyslack-package.R
\docType{package}
\name{tidyslack-package}
\alias{tidyslack}
\alias{tidyslack-package}
\title{tidyslack: Tidy Slack Conversations}
\description{
Clean and organize data retrieved from the Slack api.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jonthegeek.github.io/tidyslack/}
  \item \url{https://github.com/jonthegeek/tidyslack}
  \item Report bugs at \url{https://github.com/jonthegeek/tidyslack/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
